requirements.txt
requests==2.25.1
facebook_sdk==3.1.0
yahoo_finance==1.4.0
numpy==1.19.5
tweepy==3.10.0
yahoo_fin==0.8.7
pandas==1.2.1
matplotlib==3.3.4
pandas_datareader==0.9.0
instabot==0.117.0
python_wordpress_xmlrpc==2.3
Pillow==8.3.2
python_twitter==3.5
